pyproject.toml
[tool.poetry]
name = "zarea-de-risco"
version = "0.1.0"
description = "a initial poetry project for Zarea de Risco"
authors = ["welli7ngton <github.com/welli7ngton>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aniso8601 = "9.0.1"
argon2-cffi = "23.1.0"
argon2-cffi-bindings = "21.2.0"
blinker = "1.9.0"
cachecontrol = "0.14.1"
certifi = "2024.8.30"
cffi = "1.17.1"
charset-normalizer = "3.4.0"
click = "8.1.7"
colorama = "0.4.6"
cryptography = "44.0.0"
distlib = "0.3.9"
dulwich = "0.21.7"
fastjsonschema = "2.21.1"
flask = "3.1.0"
flask-restful = "0.3.10"
idna = "3.10"
importlib-metadata = "8.5.0"
itsdangerous = "2.2.0"
jinja2 = "3.1.4"
keyring = "24.3.1"
markupsafe = "3.0.2"
minio = "7.2.12"
more-itertools = "10.5.0"
msgpack = "1.1.0"
numpy = "2.2.0"
packaging = "24.2"
pandas = "2.2.3"
pexpect = "4.9.0"
platformdirs = "4.3.6"
psutil = "6.1.0"
ptyprocess = "0.7.0"
pyaes = "1.6.1"
pyasn1 = "0.6.1"
pycparser = "2.22"
pycryptodome = "3.21.0"
python-dateutil = "2.9.0.post0"
python-dotenv = "1.0.1"
pytz = "2024.2"
rapidfuzz = "3.11.0"
requests = "2.32.3"
requests-toolbelt = "1.0.0"
rsa = "4.9"
secretstorage = "3.3.3"
six = "1.17.0"
telethon = "1.38.1"
tomli = "2.2.1"
tqdm = "4.67.1"
tzdata = "2024.2"
urllib3 = "2.2.3"
werkzeug = "3.1.3"
zipp = "3.21.0"
zstandard = "0.23.0"
black = "^24.10.0"


[tool.poetry.group.dev.dependencies]
taskipy = "^1.14.1"
black = "^24.10.0"
flake8 = "^7.1.1"
google-auth = "^2.37.0"
google-api-python-client = "^2.156.0"
google-auth-oauthlib = "^1.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.taskipy.tasks]
# flask app
zarea_build = { cmd = "docker compose -f ./backend/docker-compose.yml up -d --build", help = "build the flask app" }
zarea_up = { cmd = "docker start zarea", help = "start flask app"}
zarea_res = { cmd = "docker stop zarea && echo restarting container... && task zarea_up", help = "restart flask app"}
zarea_down = { cmd = "docker stop zarea && echo removing container... && task zarea_rm", help = "stop and remove flask app"}
zarea_rm = { cmd = "docker container rm zarea", help = "remove flask app"}

# minio
minio = { cmd = "docker run -d -p 9000:9000 -p 9090:9090 --name minio -v /mnt/wsl/minio/data:/data -v /mnt/wsl/minio/config:/root/.minio minio/minio server /data --console-address ':9090'", help = "run minio"}
start_min = { cmd = "docker start minio", help = "start minio container" }
stop_min = { cmd = "docker stop minio", help = "stop minio container" }
rm_min = { cmd = "docker container rm minio", help = "remove minio container" }

# jobs
job_up = { cmd = "python ./job_upload_to_minio.py", help = "execute the upload file"}
job_down = { cmd = "python ./job_download_data.py.py", help = "execute the download file"}

# docker general
ps = {cmd = "docker ps -a", help = "show all the containers running or not"}
